replace NaturalDatetimeColumn with actual date & datetime on hover
