+++ Seeder/core/utils.py
@@ -1,10 +1,5 @@
 # pylint: disable=W0613
 
-from django.db import models
-from django.contrib.humanize.templatetags.humanize import naturaltime
-from django_tables2.columns.linkcolumn import BaseLinkColumn
-from django_tables2.columns.base import Column
-
 
 def merge_dicts(x, y):
     """
@@ -42,39 +37,6 @@ def dict_diff(first, second):
                 'changed': second_value
             }
     return diff
-
-
-class AbsoluteURLColumn(BaseLinkColumn):
-    """
-        Django tables2 column that renders models get_absolute_url
-    """
-
-    def render(self, record, table, value, bound_column, **kwargs):
-        """
-        This is kind of dirty, but django-tables2 uses some weirds tricks.
-        (Look at render method of LinkColumn and URLColumn, each has different
-        set of arguments!)
-
-        If this was created using accessor, value is the thing we want.
-        Otherwise value is the text to display and record is the main model.
-        """
-
-        url = (
-            value.get_absolute_url() if isinstance(value, models.Model)
-            else record.get_absolute_url()
-        )
-        return self.render_link(
-            uri=url, record=record, value=value
-        )
-
-
-class NaturalDatetimeColumn(Column):
-    """
-        Django tables2 column with human date times
-    """
-
-    def render(self, value):
-        return naturaltime(value)
 
 
 def show_toolbar(request):
