+++ Seeder/publishers/tables.py
@@ -1,7 +1,7 @@
 import django_tables2 as tables
 
 from django.utils.translation import ugettext_lazy as _
-from core.utils import AbsoluteURLColumn, NaturalDatetimeColumn
+from core.tables import AbsoluteURLColumn, NaturalDatetimeColumn
 from . import models
