+++ Seeder/source/tables.py
@@ -1,7 +1,6 @@
 import django_tables2 as tables
-
 from django.utils.translation import ugettext_lazy as _
-from core.utils import AbsoluteURLColumn, NaturalDatetimeColumn, Column
+from core.tables import AbsoluteURLColumn, NaturalDatetimeColumn
 from . import models
 
 
@@ -11,9 +10,9 @@ class SourceTable(tables.Table):
         verbose_name=_('Publisher'),
         accessor='publisher'
     )
-    owner = Column(order_by='owner_id')
-    category = Column(order_by='category_id')
-    sub_category = Column(order_by='sub_category_id')
+    owner = tables.Column(order_by='owner_id')
+    category = tables.Column(order_by='category_id')
+    sub_category = tables.Column(order_by='sub_category_id')
 
     created = NaturalDatetimeColumn(verbose_name=_('Created'))
     last_changed = NaturalDatetimeColumn(verbose_name=_('Last changed'))
